Added Spacecraft & Missiles
Spacecraft are able to move with A, W, S, D keys and arrow keys. 1
missile can be fired from crusier spacecraft, to fire press F. Missiles
will be worked on to be fired multiple times and for both spacecraft.
Missile bugs will also be fixed and sound effects will be added.